requirements.txt
attrs==22.1.0
beaker-pyteal==0.4.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
docstring-parser==0.14.1
exceptiongroup==1.0.4
iniconfig==1.1.1
msgpack==1.0.4
packaging==22.0
pluggy==1.0.0
py-algorand-sdk==1.20.2
pycparser==2.21
pycryptodomex==3.16.0
PyNaCl==1.5.0
pyteal==0.20.1
pytest==7.2.0
semantic-version==2.10.0
tomli==2.0.1
